fix: 로그아웃 로직 헤더로 이동
