+++ src/layouts/Header/UserDropdown.tsx
@@ -4,11 +4,6 @@ import Link from "next/link";
 import DropDownMenu from "@/components/common/Dropdown/Menu";
 import DropDownItem from "@/components/common/Dropdown/Item";
 import useClickOutside from "@/hooks/useClickOutside";
-import { useSignOut } from "@/api/auth/auth.query";
-import { useModalStore } from "@/stores/modalStore";
-import LogoutModal from "@/components/common/Modal/LogoutModal";
-import { useToastStore } from "@/stores/toastStore";
-import { useRouter } from "next/navigation";
 
 interface UserDropdownProps {
   userName: string;
@@ -16,6 +11,7 @@ interface UserDropdownProps {
   isOpen: boolean;
   onToggle: () => void;
   onClose: () => void;
+  onLogoutClick: () => void;
 }
 
 export default function UserDropdown({
@@ -24,24 +20,9 @@ export default function UserDropdown({
   isOpen,
   onToggle,
   onClose,
+  onLogoutClick,
 }: UserDropdownProps) {
   const ref = useClickOutside(onClose);
-  const signOutMutation = useSignOut();
-  const { openModal, closeModal } = useModalStore();
-  const { showToast } = useToastStore();
-  const router = useRouter();
-
-  const handleSignOut = () => {
-    signOutMutation.mutate(undefined, {
-      onSuccess: () => {
-        {
-          closeModal();
-          showToast("로그아웃 되었습니다.", "success");
-          router.replace("/");
-        }
-      },
-    });
-  };
 
   return (
     <div ref={ref} className="relative">
@@ -82,12 +63,8 @@ export default function UserDropdown({
             계정 설정
           </Link>
         </DropDownItem>
-        <DropDownItem onClick={() => openModal("logout")}>
-          로그아웃
-        </DropDownItem>
+        <DropDownItem onClick={onLogoutClick}>로그아웃</DropDownItem>
       </DropDownMenu>
-
-      <LogoutModal onConfirm={handleSignOut} />
     </div>
   );
 }
